+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # dev
 
 * Added support for tokenizer action when end of input stream is reached
+* Token actions are now always called exactly once for each individual token
 
 # v0.1.0 (2019-06-27)